# public/docs/openapi/openAPI.yml
openapi: 3.0.0
servers:
  - url: https://paywithpretendpointsapi.onrender.com
info:
  version: 1.0.0
  title: Pay With Pretend Points API
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/loyalty/{memberId}/transactions/{transactionId}:
    get:
      tags:
        - Member Transactions
      summary: Get a specific transaction for a member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: transactionId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      memberId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerRefId:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: INV-20234
                      reference:
                        type: string
                        format: uuid
                        example: 123e4567-e89b-12d3-a456-426614174001
                      amount:
                        type: integer
                        example: 100
                      note:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: This is a note about the transaction.
                      status:
                        type: string
                        enum: *a1
                        default: PENDING
                        example: PENDING
                      type:
                        type: string
                        enum: *a2
                        default: PAYMENT
                        example: PAYMENT
                      transactedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - partnerId
                      - memberId
                      - reference
                      - amount
                      - transactedAt
                      - createdAt
                      - updatedAt
                required:
                  - transaction
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    patch:
      tags:
        - Member Transactions
      summary: Update a specific transaction for a member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: transactionId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                partnerRefId:
                  type: string
                  pattern: ^[\w\s.,!?'"()-]+$
                  example: INV-20234
                amount:
                  type: integer
                  example: 100
                note:
                  type: string
                  pattern: ^[\w\s.,!?'"()-]+$
                  example: This is a note about the transaction.
                type:
                  type: string
                  enum: *a2
                  default: PAYMENT
                  example: PAYMENT
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      memberId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerRefId:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: INV-20234
                      reference:
                        type: string
                        format: uuid
                        example: 123e4567-e89b-12d3-a456-426614174001
                      amount:
                        type: integer
                        example: 100
                      note:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: This is a note about the transaction.
                      status:
                        type: string
                        enum: *a1
                        default: PENDING
                        example: PENDING
                      type:
                        type: string
                        enum: *a2
                        default: PAYMENT
                        example: PAYMENT
                      transactedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - partnerId
                      - memberId
                      - reference
                      - amount
                      - transactedAt
                      - createdAt
                      - updatedAt
                required:
                  - transaction
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    delete:
      tags:
        - Member Transactions
      summary: Reverses a transaction for a member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: transactionId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 1
                required:
                  - count
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
  /api/v1/loyalty/{memberId}/transactions:
    get:
      tags:
        - Member Transactions
      summary: Get all transactions for a member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          minimum: 0
                          exclusiveMinimum: true
                          example: 1
                        partnerId:
                          type: integer
                          minimum: 0
                          exclusiveMinimum: true
                          example: 1
                        memberId:
                          type: integer
                          minimum: 0
                          exclusiveMinimum: true
                          example: 1
                        partnerRefId:
                          type: string
                          pattern: ^[\w\s.,!?'"()-]+$
                          example: INV-20234
                        reference:
                          type: string
                          format: uuid
                          example: 123e4567-e89b-12d3-a456-426614174001
                        amount:
                          type: integer
                          example: 100
                        note:
                          type: string
                          pattern: ^[\w\s.,!?'"()-]+$
                          example: This is a note about the transaction.
                        status:
                          type: string
                          enum: *a1
                          default: PENDING
                          example: PENDING
                        type:
                          type: string
                          enum: *a2
                          default: PAYMENT
                          example: PAYMENT
                        transactedAt:
                          type: string
                          example: 2024-09-01T01:03:43.004Z
                        createdAt:
                          type: string
                          example: 2024-09-01T01:03:43.004Z
                        updatedAt:
                          type: string
                          example: 2024-09-01T01:03:43.004Z
                      required:
                        - id
                        - partnerId
                        - memberId
                        - reference
                        - amount
                        - transactedAt
                        - createdAt
                        - updatedAt
                required:
                  - transactions
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
    post:
      tags:
        - Member Transactions
      summary: Create a new transaction for a member.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 0
            exclusiveMinimum: true
            example: 1
          required: true
          name: memberId
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                partnerRefId:
                  type: string
                  pattern: ^[\w\s.,!?'"()-]+$
                  example: INV-20234
                amount:
                  type: integer
                  example: 100
                note:
                  type: string
                  pattern: ^[\w\s.,!?'"()-]+$
                  example: This is a note about the transaction.
                type:
                  type: string
                  enum: *a2
                  default: PAYMENT
                  example: PAYMENT
              required:
                - amount
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction:
                    type: object
                    properties:
                      id:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      memberId:
                        type: integer
                        minimum: 0
                        exclusiveMinimum: true
                        example: 1
                      partnerRefId:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: INV-20234
                      reference:
                        type: string
                        format: uuid
                        example: 123e4567-e89b-12d3-a456-426614174001
                      amount:
                        type: integer
                        example: 100
                      note:
                        type: string
                        pattern: ^[\w\s.,!?'"()-]+$
                        example: This is a note about the transaction.
                      status:
                        type: string
                        enum: *a1
                        default: PENDING
                        example: PENDING
                      type:
                        type: string
                        enum: *a2
                        default: PAYMENT
                        example: PAYMENT
                      transactedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      createdAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                      updatedAt:
                        type: string
                        example: 2024-09-01T01:03:43.004Z
                    required:
                      - id
                      - partnerId
                      - memberId
                      - reference
                      - amount
                      - transactedAt
                      - createdAt
                      - updatedAt
                required:
                  - transaction
        '204':
          description: No Content
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
...